refactor: API 모듈 구조 리팩토링
   - 상품(랭킹) 및 테마 API 모듈 분리
   - GiftRankingSection, GiftThemeSection 컴포넌트의 API 연동 코드 수정
